requirements.txt
Flask==2.3.2
flask-cors==3.0.10
torch==2.2.0
torchvision==0.17.1
Pillow==10.1.0
geopandas==1.2.3
shapely==2.1.3
matplotlib==3.7.1
